+++ app/types/declarations.d.ts
@@ -1,6 +1,5 @@
 declare module 'json-canonicalize'
 declare module 'rn-animated-ellipsis'
-declare module 'react-native-html-to-pdf'
 declare module 'validator'
 declare module '@microsoft/msrcrypto'